+++ twitchio/errors.py
@@ -21,7 +21,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
 FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 DEALINGS IN THE SOFTWARE.
 """
-from typing import Optional
+from typing import Any, Optional
 
 
 __all__ = (
@@ -67,9 +67,16 @@ class NoToken(TwitchIOException):
 
 class HTTPException(TwitchIOException):
 
-    def __init__(self, reason: Optional[str] = None, status: Optional[int] = None):
+    def __init__(self, message: str,
+                 *,
+                 reason: Optional[str] = None,
+                 status: Optional[int] = None,
+                 extra: Optional[Any] = None):
+
+        self.message = message
         self.reason = reason
         self.status = status
+        self.extra = extra
 
 
 class Unauthorized(HTTPException):