+++ docs/changelog.rst
@@ -8,10 +8,10 @@
 - TwitchIO
     - Loosen aiohttp requirements to allow 3.8.1
     - :class:`Stream` was missing from ``__all__``. It is now available in the twitchio namespace.
-    - Added ``.status`` and ``.reason`` to :class:`HTTPException`
+    - Added ``.status``, ``.reason`` and ``.extra`` to :class:`HTTPException`
   
 - ext.pubsub
-    - Add channel subscription pubsub model
+    - Add channel subscription pubsub model.
 
 2.1.5
 =====